main function and report

+++ includes/merge_sort.hpp
@@ -1,4 +1,4 @@
-void merge(int arr[], int l, int m, int r, int &invCnt) {
+void merge(int arr[], int l, int m, int r, long long &invCnt) {
     int n1 = m - l + 1;
     int n2 = r - m;
 
@@ -36,7 +36,7 @@ void merge(int arr[], int l, int m, int r, int &invCnt) {
     }
 }
 
-void mergeSort(int arr[], int l, int r, int &invCnt) {
+void mergeSort(int arr[], int l, int r, long long &invCnt) {
     if (l >= r) {
         return;
     }
